fix(workbooks): preserve calculation and XLSX round-trip contracts
Integrate imported formula modes and reference semantics, deferred calculation state, visibility-aware aggregation, typed icon filters, pivot totals, metadata preservation, and validated drawing ownership. Preserve cached values on ordinary export.

Checkpoint while final verification remains open. The completed corpus exported all 82 cases in both modes without execution errors; the continuing runtime gate found INDEX bounds and deferred-import scheduling defects whose fixes are included here and awaiting final reruns.

+++ domain-types/src/domain/filter/conversion.rs
@@ -126,7 +126,7 @@ fn ooxml_filter_type_to_column_filter(ft: &OoxmlFilterType) -> ColumnFilter {
         },
         OoxmlFilterType::Icon { icon_set, icon_id } => ColumnFilter::Icon {
             icon_set_name: icon_set.clone().unwrap_or_default(),
-            icon_index: *icon_id as u8,
+            icon_index: *icon_id,
         },
     }
 }
@@ -338,7 +338,7 @@ pub fn column_filter_to_ooxml_filter_type(cf: &ColumnFilter) -> OoxmlFilterType 
             } else {
                 Some(icon_set_name.clone())
             },
-            icon_id: *icon_index as u32,
+            icon_id: *icon_index,
         },
     }
 }
